infra/requirements.txt
-i https://pypi.org/simple
attrs==24.2.0; python_version >= '3.7'
aws-cdk-lib==2.158.0; python_version ~= '3.8'
aws-cdk.asset-awscli-v1==2.2.202; python_version ~= '3.8'
aws-cdk.asset-kubectl-v20==2.1.2; python_version ~= '3.7'
aws-cdk.asset-node-proxy-agent-v6==2.1.0; python_version ~= '3.8'
aws-cdk.cloud-assembly-schema==36.0.25; python_version ~= '3.8'
cattrs==23.2.3; python_version >= '3.8'
constructs==10.3.0; python_version ~= '3.7'
exceptiongroup==1.2.2; python_version < '3.11'
importlib-resources==6.4.5; python_version >= '3.8'
jsii==1.103.1; python_version ~= '3.8'
publication==0.0.3
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typeguard==2.13.3; python_full_version >= '3.5.3'
typing-extensions==4.12.2; python_version >= '3.8'
zipp==3.20.2; python_version < '3.10'
